python-backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pillow==11.0.0
numpy==2.2.6
picamera2
